Implementation of the deletion dependency

+++ src/main/java/com/ntarasov/store/admin/service/AdminApiService.java
@@ -28,7 +28,8 @@ public class AdminApiService {
     private final MongoTemplate mongoTemplate;
 
     //<---------------------------------ПОИСК ПО ID------------------------------------------------->
-    public Optional<AdminDoc> findById(ObjectId id) {
+    public Optional<AdminDoc> findById(ObjectId id) throws AdminNotExistException {
+        if(adminRepository.findById(id).isEmpty()) throw new AdminNotExistException();
         return adminRepository.findById(id);
     }
 
@@ -68,13 +69,15 @@ public class AdminApiService {
         Optional<AdminDoc> adminDocOptional = adminRepository.findById(request.getId());
         if (adminDocOptional.isEmpty()) throw new AdminNotExistException();
         AdminDoc adminDoc = AdminMapping.getInstance().getRequest().convert(request);
+        adminDoc.setEmail(request.getEmail());
         adminDoc.setId(request.getId());
         adminRepository.save(adminDoc);
         return adminDoc;
     }
 
     //<---------------------------------УДАЛЕНИЕ------------------------------------------------->
-    public void delete(ObjectId id) {
+    public void delete(ObjectId id) throws AdminNotExistException {
+        if(adminRepository.findById(id).isEmpty()) throw new AdminNotExistException();
         adminRepository.deleteById(id);
     }
 }
